실수 계산및 출력ã 기능 추가

+++ src/com/example/calculator/level3/OperatorType.java
@@ -3,25 +3,25 @@ package com.example.calculator.level3;
 public enum OperatorType {
     ADD('+') {
         @Override
-        public int calculate(int a, int b) {
+        public double calculate(double a, double b) {
             return a + b;
         }
     },
     SUB('-') {
         @Override
-        public int calculate(int a, int b) {
+        public double calculate(double a, double b) {
             return a - b;
         }
     },
     MUL('*') {
         @Override
-        public int calculate(int a, int b) {
+        public double calculate(double a, double b) {
             return a * b;
         }
     },
     DIV('/') {
         @Override
-        public int calculate(int a, int b) {
+        public double calculate(double a, double b) {
             if (b == 0) {
                 throw new ArithmeticException("나눗셈 연산에서 분모(두번째 정수)에 0이 입력될 수 없습니다.");
             }
@@ -42,7 +42,7 @@ public enum OperatorType {
     }
 
     // 추상 메서드 선언
-    public abstract int calculate(int a, int b);
+    public abstract double calculate(double a, double b);
 
     // Symbol로 Enum 상수 찾기
     public static OperatorType fromSymbol(char symbol) {
